improve clean scripts
