+++ services/apps/script_executor_worker/src/activities.ts
@@ -1,4 +1,4 @@
-import { excludeEntityFromCleanup, hasActivityRecords } from './activities/cleanup/helpers'
+import { excludeEntityFromCleanup } from './activities/cleanup/exlcudeEntity'
 import { deleteMember, getMembersToCleanup } from './activities/cleanup/member'
 import {
   deleteOrganization,
@@ -6,6 +6,7 @@ import {
   queueOrgForAggComputation,
 } from './activities/cleanup/organization'
 import {
+  hasActivityRecords,
   mergeMembers,
   mergeOrganizations,
   unmergeMembers,
