+++ services/apps/script_executor_worker/src/activities/common.ts
@@ -1,6 +1,8 @@
 import axios from 'axios'
 
 import { findOrganizationSegments } from '@crowd/data-access-layer/src/old/apps/entity_merging_worker'
+import ActivityRepository from '@crowd/data-access-layer/src/old/apps/script_executor_worker/activity.repo'
+import { EntityType } from '@crowd/data-access-layer/src/old/apps/script_executor_worker/types'
 import {
   IMemberIdentity,
   IMemberUnmergeBackup,
@@ -133,4 +135,21 @@ export function timeout(ms: number, workflowId: string): Promise<void> {
       reject(new Error(`Timeout waiting for workflow ${workflowId} to finish`))
     }, ms)
   })
+}
+
+export async function hasActivityRecords(
+  entityId: string,
+  entityType: EntityType,
+): Promise<boolean> {
+  try {
+    const activityRepo = new ActivityRepository(
+      svc.postgres.reader.connection(),
+      svc.log,
+      svc.questdbSQL,
+    )
+    return activityRepo.hasActivitiesInQuestDb(entityId, entityType)
+  } catch (error) {
+    svc.log.error(error, 'Error checking if entity has activities in questDb!')
+    throw error
+  }
 }
