+++ services/apps/script_executor_worker/src/workflows.ts
@@ -6,8 +6,8 @@ import { findAndMergeMembersWithSamePlatformIdentitiesDifferentCapitalization } 
 import { findAndMergeMembersWithSameVerifiedEmailsInDifferentPlatforms } from './workflows/findAndMergeMembersWithSameVerifiedEmailsInDifferentPlatforms'
 import { fixOrgIdentitiesWithWrongUrls } from './workflows/fixOrgIdentitiesWithWrongUrls'
 import { populateActivityRelations } from './workflows/populateActivityRelations'
-import { syncMembers } from './workflows/syncMembers'
-import { syncOrganizations } from './workflows/syncOrganizations'
+import { syncMembers } from './workflows/sync/members'
+import { syncOrganizations } from './workflows/sync/organizations'
 
 export {
   findAndMergeMembersWithSameVerifiedEmailsInDifferentPlatforms,
