+++ services/apps/script_executor_worker/src/workflows/cleanup/organizations.ts
@@ -2,16 +2,16 @@ import { continueAsNew, proxyActivities } from '@temporalio/workflow'
 
 import { EntityType } from '@crowd/data-access-layer/src/old/apps/script_executor_worker/types'
 
-import * as cleanupHelpers from '../../activities/cleanup/helpers'
-import * as activities from '../../activities/cleanup/organization'
+import * as activities from '../../activities'
 import { ICleanupArgs } from '../../types'
 
-const activity = proxyActivities<typeof activities>({
-  startToCloseTimeout: '30 minutes',
-  retry: { maximumAttempts: 3, backoffCoefficient: 3 },
-})
-
-const cleanupHelper = proxyActivities<typeof cleanupHelpers>({
+const {
+  getOrganizationsToCleanup,
+  deleteOrganization,
+  queueOrgForAggComputation,
+  hasActivityRecords,
+  excludeEntityFromCleanup,
+} = proxyActivities<typeof activities>({
   startToCloseTimeout: '30 minutes',
   retry: { maximumAttempts: 3, backoffCoefficient: 3 },
 })
@@ -19,7 +19,7 @@ const cleanupHelper = proxyActivities<typeof cleanupHelpers>({
 export async function cleanupOrganizations(args: ICleanupArgs): Promise<void> {
   const BATCH_SIZE = args.batchSize ?? 100
 
-  const organizationIds = await activity.getOrganizationsToCleanup(BATCH_SIZE)
+  const organizationIds = await getOrganizationsToCleanup(BATCH_SIZE)
 
   if (organizationIds.length === 0) {
     console.log('No more organizations to cleanup!')
@@ -32,17 +32,17 @@ export async function cleanupOrganizations(args: ICleanupArgs): Promise<void> {
     const chunk = organizationIds.slice(i, i + CHUNK_SIZE)
 
     const cleanupTasks = chunk.map(async (orgId) => {
-      const isInQuestDb = await cleanupHelper.hasActivityRecords(orgId, EntityType.ORGANIZATION)
+      const isInQuestDb = await hasActivityRecords(orgId, EntityType.ORGANIZATION)
 
       if (isInQuestDb) {
         console.log(`Organization ${orgId} is in QuestDB, skipping!`)
-        return cleanupHelper.excludeEntityFromCleanup(orgId, EntityType.ORGANIZATION)
+        return excludeEntityFromCleanup(orgId, EntityType.ORGANIZATION)
       }
 
       console.log(`Deleting organization ${orgId} from database!`)
+      await deleteOrganization(orgId)
 
-      await activity.deleteOrganization(orgId)
-      return activity.queueOrgForAggComputation(orgId)
+      return queueOrgForAggComputation(orgId)
     })
 
     await Promise.all(cleanupTasks)
